requirements.txt
bert_score==0.3.13
Distance==0.1.3
ltp
# numpy==1.22.3
OpenCC
# pandas==1.4.2
pypinyin==0.49.0
python_Levenshtein==0.25.1
pytorch_lightning
# scikit_learn==1.1.0
# torch==1.11.0+cu113
# torchmetrics==0.8.2
# tqdm==4.64.0
# transformers==4.35.2
